# openapi.yaml
openapi: 3.0.3
info:
  title: devos.ing Server API
  version: 0.1.0
  description: Implemented HTTP routes for the Express-backed server.
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/notifications:
    post:
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: Notification accepted.
        '400':
          $ref: '#/components/responses/TypedError'
  /api/notifications/email:
    post:
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Notification sent.
        '400':
          $ref: '#/components/responses/TypedError'
components:
...
  responses:
    TypedError:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TypedError'
  schemas:
    TypedError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...